Separate rnp_ctx_t-related declarations and functions to stream-ctx.h/stream-ctx.cpp
