+++ include/rnp/rnp_types.h
@@ -86,90 +86,4 @@ typedef struct rnp_params_t {
     pgp_password_provider_t password_provider;
 } rnp_params_t;
 
-typedef struct rnp_symmetric_pass_info_t {
-    pgp_s2k_t      s2k;
-    pgp_symm_alg_t s2k_cipher;
-    uint8_t        key[PGP_MAX_KEY_SIZE];
-} rnp_symmetric_pass_info_t;
-
-typedef enum rnp_operation_t {
-    RNP_OP_UNKNOWN = 0,
-    RNP_OP_DECRYPT_VERIFY = 1,
-    RNP_OP_ENCRYPT_SIGN = 2,
-    RNP_OP_ARMOR = 3
-} rnp_operation_t;
-
-/* signature info structure */
-typedef struct rnp_signer_info_t {
-    pgp_key_t *    key;
-    pgp_hash_alg_t halg;
-    int64_t        sigcreate;
-    uint64_t       sigexpire;
-} rnp_signer_info_t;
-
-/** rnp operation context : contains configuration data about the currently ongoing operation.
- *
- *  Common fields which make sense for every operation:
- *  - overwrite : silently overwrite output file if exists
- *  - armor : except cleartext signing, which outputs text in clear and always armor signature,
- *    this controls whether output is armored (base64-encoded). For armor/dearmor operation it
- *    controls the direction of the conversion (true means enarmor, false - dearmor),
- *  - rng : random number generator
- *  - operation : current operation type
- *
- *  For operations with OpenPGP embedded data (i.e. encrypted data and attached signatures):
- *  - filename, filemtime : to specify information about the contents of literal data packet
- *  - zalg, zlevel : compression algorithm and level, zlevel = 0 to disable compression
- *
- *  For encryption operation (including encrypt-and-sign):
- *  - halg : hash algorithm used during key derivation for password-based encryption
- *  - ealg, aalg, abits : symmetric encryption algorithm and AEAD parameters if used
- *  - recipients : list of key ids used to encrypt data to
- *  - passwords : list of passwords used for password-based encryption
- *  - filename, filemtime, zalg, zlevel : see previous
- *
- *  For signing of any kind (attached, detached, cleartext):
- *  - clearsign, detached : controls kind of the signed data. Both are mutually-exclusive.
- *    If both are false then attached signing is used.
- *  - halg : hash algorithm used to calculate signature(s)
- *  - signers : list of rnp_signer_info_t structures describing signing key and parameters
- *  - sigcreate, sigexpire : default signature(s) creation and expiration times
- *  - filename, filemtime, zalg, zlevel : only for attached signatures, see previous
- *
- *  For data decryption and/or verification there is not much of fields:
- *  - on_signatures: callback, called when signature verification information is available.
- *    If we have just encrypted data then it will not be called.
- *  - sig_cb_param: parameter to be passed to on_signatures callback.
- *  - discard: dicard the output data (i.e. just decrypt and/or verify signatures)
- *
- *  For enarmor/dearmor:
- *  - armortype: type of the armor headers (message, key, whatever else)
- */
-
-typedef struct rnp_ctx_t {
-    char *          filename;      /* name of the input file to store in literal data packet */
-    int64_t         filemtime;     /* file modification time to store in literal data packet */
-    int64_t         sigcreate;     /* signature creation time */
-    uint64_t        sigexpire;     /* signature expiration time */
-    bool            clearsign;     /* cleartext signature */
-    bool            detached;      /* detached signature */
-    pgp_hash_alg_t  halg;          /* hash algorithm */
-    pgp_symm_alg_t  ealg;          /* encryption algorithm */
-    int             zalg;          /* compression algorithm used */
-    int             zlevel;        /* compression level */
-    pgp_aead_alg_t  aalg;          /* non-zero to use AEAD */
-    int             abits;         /* AEAD chunk bits */
-    bool            overwrite;     /* allow to overwrite output file if exists */
-    bool            armor;         /* whether to use ASCII armor on output */
-    list            recipients;    /* recipients of the encrypted message */
-    list            passwords;     /* list of rnp_symmetric_pass_info_t */
-    list            signers;       /* list of rnp_signer_info_t structures */
-    unsigned        armortype;     /* type of the armored message, used in enarmor command */
-    bool            discard;       /* discard the output */
-    void *          on_signatures; /* handler for signed messages */
-    void *          sig_cb_param;  /* callback data passed to on_signatures */
-    rng_t *         rng;           /* pointer to rng_t */
-    rnp_operation_t operation;     /* current operation type */
-} rnp_ctx_t;
-
 #endif // __RNP_TYPES__
