+++ src/apps/packet-dumper/redumper.cpp
@@ -1,6 +1,7 @@
 #include <stdio.h>
 #include <unistd.h> /* getopt() */
 #include <getopt.h>
+#include <librepgp/stream-ctx.h>
 #include <librepgp/stream-dump.h>
 #include "utils.h"
 #include <rnp/rnp.h> /* rnp_t, rnp_ctx_t et. all */
