+++ src/lib/rnp2.cpp
@@ -33,6 +33,7 @@
 #include <assert.h>
 #include <json_object.h>
 #include <librepgp/packet-show.h>
+#include <librepgp/stream-ctx.h>
 #include <librepgp/stream-common.h>
 #include <librepgp/stream-armor.h>
 #include <librepgp/stream-parse.h>
