+++ src/rnp/rnp.cpp
@@ -45,6 +45,7 @@
 #include <regex.h>
 #include <rnp/rnp.h>
 #include <rnp/rnp_sdk.h>
+#include <librepgp/stream-ctx.h>
 #include <librepgp/stream-parse.h>
 #include <librepgp/stream-armor.h>
 #include <librepgp/stream-packet.h>
