+++ src/tests/key-unlock.cpp
@@ -25,6 +25,7 @@
  */
 
 #include "../librekey/key_store_pgp.h"
+#include "../librepgp/stream-ctx.h"
 #include "pgp-key.h"
 
 #include "rnp_tests.h"
